requirements.txt
click
kubernetes
docker
python-json-logger
pyyaml
colorlog
colorama
